requirements.txt
beautifulsoup4==4.12.3
bs4==0.0.2
cached-property==1.5.2
defusedxml==0.7.1
entrypoints==0.4
exceptiongroup==1.2.0
jupyter==1.0.0
numpy==1.26.3
overrides==7.7.0
packaging==23.2
pandas==2.1.4
pillow==10.2.0
ptyprocess==0.7.0
pyzbar==0.1.9
pyzmq==25.1.2
referencing==0.33.0
requests==2.31.0
stack-data==0.6.2
stqdm==0.0.5
streamlit==1.32.0
streamlit-aggrid==0.3.4.post3
streamlit-image-select==0.6.0
tqdm==4.66.2


